requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiopg==1.4.0
    # via apprendre (pyproject.toml)
alembic==1.16.5
    # via apprendre (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via aiopg
click==8.3.0
    # via uvicorn
dependency-injector==4.48.2
    # via apprendre (pyproject.toml)
fastapi==0.118.0
    # via apprendre (pyproject.toml)
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
iniconfig==2.1.0
    # via pytest
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
psycopg==3.2.10
    # via apprendre (pyproject.toml)
psycopg-binary==3.2.10
    # via psycopg
psycopg2-binary==2.9.10
    # via
    #   apprendre (pyproject.toml)
    #   aiopg
pydantic==2.11.9
    # via
    #   apprendre (pyproject.toml)
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via apprendre (pyproject.toml)
pydocstyle==6.3.0
    # via apprendre (pyproject.toml)
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via apprendre (pyproject.toml)
pytest==8.4.2
    # via apprendre (pyproject.toml)
python-dotenv==1.1.1
    # via
    #   apprendre (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
snowballstemmer==3.0.1
    # via pydocstyle
sqlalchemy==2.0.43
    # via
    #   apprendre (pyproject.toml)
    #   alembic
    #   sqlmodel
sqlmodel==0.0.25
    # via apprendre (pyproject.toml)
starlette==0.48.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.37.0
    # via apprendre (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
